+++ src/bin/server/copier.rs
@@ -3,6 +3,7 @@
 use log::{debug, error, info, warn};
 
 use std::collections::{HashMap, LinkedList};
+use std::process::{Child, Command, Output, Stdio};
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, Instant};
 
@@ -12,8 +13,6 @@ const RSYNC_TIMEOUT: u64 = 60;
 const RSYNC_IO_TIMEOUT: u64 = 60; // seconds
 /// The number of times to retry before giving up.
 const RETRIES: usize = 3;
-/// A little log directory for rsync... not ideal, but better than nothing.
-const RSYNC_LOG_PATH: &str = "/tmp/jobserver-rsync.log";
 /// The max number of concurrent copy jobs.
 const MAX_CONCURRENCY: usize = 2;
 
@@ -36,8 +35,16 @@ struct CopierThreadState {
     incoming: Arc<Mutex<CopierThreadQueue>>,
     /// Copy tasks that are currently running.
     ongoing: HashMap<u64, ResultsInfo>,
-    /// Used to notify the worker thread of completion. `true` indicates success; `false` failure.
-    copying_flags: Arc<Mutex<HashMap<u64, bool>>>,
+    /// Used to notify the worker thread of completion.
+    copying_flags: Arc<Mutex<HashMap<u64, CopierThreadResult>>>,
+}
+
+/// Possible outcomes for copying of results.
+#[derive(Debug)]
+pub enum CopierThreadResult {
+    Success,
+    SshUnknownHostKey,
+    OtherFailure,
 }
 
 /// Info about a single copy job.
@@ -52,7 +59,7 @@ struct ResultsInfo {
     /// The path on the host to copy to.
     to: String,
     /// The process currently doing the copy.
-    process: std::process::Child,
+    process: Child,
     /// The time when the process was started.
     start_time: Instant,
     /// The current attempt number.
@@ -74,7 +81,7 @@ impl std::cmp::Eq for ResultsInfo {}
 
 /// Start the copier thread and return a handle.
 pub fn init(
-    copying_flags: Arc<Mutex<HashMap<u64, bool>>>,
+    copying_flags: Arc<Mutex<HashMap<u64, CopierThreadResult>>>,
 ) -> (std::thread::JoinHandle<()>, Arc<Mutex<CopierThreadQueue>>) {
     let incoming = Arc::new(Mutex::new(LinkedList::new()));
     let state = CopierThreadState {
@@ -138,7 +145,11 @@ fn copier_thread(mut state: CopierThreadState) {
             match results_info.process.try_wait() {
                 Ok(Some(exit)) if exit.success() => {
                     info!("Finished copying results for job {}.", jid,);
-                    state.copying_flags.lock().unwrap().insert(*jid, true);
+                    state
+                        .copying_flags
+                        .lock()
+                        .unwrap()
+                        .insert(*jid, CopierThreadResult::Success);
                     to_remove.push(*jid);
                 }
                 Ok(Some(exit)) => {
@@ -146,6 +157,41 @@ fn copier_thread(mut state: CopierThreadState) {
                         "Copying results for job {} failed with exit code {} (attempt {}).",
                         jid, exit, results_info.attempt
                     );
+
+                    // Check for unknown host key SSH error and report a special error for that.
+                    // First, we can do a quick check: if SSH experienced an error, we will get
+                    // exit code 255. If that is the error code, we can use ssh-keygen to check if
+                    // there is a known fingerprint.
+                    if let Some(255) = exit.code() {
+                        let fingerprint_found = Command::new("ssh-keygen")
+                            .arg("-F")
+                            .arg(&results_info.machine)
+                            .stdout(Stdio::piped())
+                            .stderr(Stdio::piped())
+                            .output();
+
+                        // If there is a failure here, we will just move on and let the normal
+                        // retry mechanism handle it.
+                        if let Ok(Output { status, .. }) = fingerprint_found {
+                            if let Some(1) = status.code() {
+                                // Unknown host error. Report the error and move on. Retrying is
+                                // not useful.
+                                error!(
+                                    "Error copy results for job {}: SSH host key verification \
+                                     failed. Add the host to known_hosts.",
+                                    jid
+                                );
+
+                                state
+                                    .copying_flags
+                                    .lock()
+                                    .unwrap()
+                                    .insert(*jid, CopierThreadResult::SshUnknownHostKey);
+                                to_remove.push(*jid);
+                                continue;
+                            }
+                        }
+                    }
 
                     // Maybe retry.
                     if results_info.attempt < RETRIES {
@@ -161,7 +207,11 @@ fn copier_thread(mut state: CopierThreadState) {
                             "Copying results for job {} failed after {} attempts.",
                             jid, RETRIES
                         );
-                        state.copying_flags.lock().unwrap().insert(*jid, false);
+                        state
+                            .copying_flags
+                            .lock()
+                            .unwrap()
+                            .insert(*jid, CopierThreadResult::OtherFailure);
                     }
                     to_remove.push(*jid);
                 }
@@ -182,7 +232,11 @@ fn copier_thread(mut state: CopierThreadState) {
                             "Copying results for job {} failed after {} attempts.",
                             jid, RETRIES
                         );
-                        state.copying_flags.lock().unwrap().insert(*jid, false);
+                        state
+                            .copying_flags
+                            .lock()
+                            .unwrap()
+                            .insert(*jid, CopierThreadResult::OtherFailure);
                     }
                     to_remove.push(*jid);
                 }
@@ -206,7 +260,11 @@ fn copier_thread(mut state: CopierThreadState) {
                                 "Copying results for job {} failed after {} attempts.",
                                 jid, RETRIES
                             );
-                            state.copying_flags.lock().unwrap().insert(*jid, false);
+                            state
+                                .copying_flags
+                                .lock()
+                                .unwrap()
+                                .insert(*jid, CopierThreadResult::OtherFailure);
                             to_remove.push(*jid);
                         }
                     } else {
@@ -248,21 +306,22 @@ fn start_copy(
         .truncate(false)
         .create(true)
         .write(true)
-        .open(RSYNC_LOG_PATH)
+        .open(copier_log_fname(jid))
         .expect("Unable to open rsync log file.");
     let log_err = log.try_clone().expect("Unable to open rsync log err file.");
 
-    // Sometimes the command will hang spuriously. So we give it a timeout and
-    // restart if needed.
-    let mut cmd = std::process::Command::new("rsync");
+    // SSH will hang waiting for StrictHostKeyChecking to get user input confirming the key
+    // fingerprint. Instead, we tell SSH to just fail fast. We can then check for the error and
+    // report it properly to the user.
+    let mut cmd = Command::new("rsync");
     let cmd = cmd
-        .arg("-vzP")
-        .arg("--rsh=ssh")
+        .arg("-vvzP")
+        .args(&["-e", r#""ssh -o StrictHostKeyChecking=yes""#])
         .arg(&format!("--timeout={}", RSYNC_IO_TIMEOUT))
         .arg(&format!("{}:{}*", machine_ip, from))
         .arg(&to)
-        .stdout(std::process::Stdio::from(log))
-        .stderr(std::process::Stdio::from(log_err));
+        .stdout(Stdio::from(log))
+        .stderr(Stdio::from(log_err));
 
     debug!("{:?}", cmd);
 
@@ -277,4 +336,9 @@ fn start_copy(
         start_time: std::time::Instant::now(),
         attempt,
     })
+}
+
+fn copier_log_fname(jid: u64) -> String {
+    const RSYNC_LOG_PATH: &str = "/tmp/jobserver-rsync";
+    format!("{}-{}.log", RSYNC_LOG_PATH, jid)
 }